+++ packages/pages/src/common/src/template/internal/types.ts
@@ -59,7 +59,10 @@ export interface TemplateConfigInternal {
   streamId?: string;
   /** The stream configuration used by the template */
   stream?: StreamInternal;
-  /** The specific fields to add additional language options to based on the stream's localization */
+  /**
+   * The specific fields to add additional language options to based on the stream's localization.
+   * @deprecated field will be unsupported in the future
+   */
   alternateLanguageFields?: string[];
   /** The name of the onUrlChange function to use. */
   onUrlChange?: string;
@@ -69,6 +72,8 @@ export interface TemplateConfigInternal {
   pageUrlField?: string;
   /** The field to use as the slug for dynamic dev mode */
   slugField?: string;
+  /** The type of template */
+  templateType: "entity" | "static";
 }
 
 /**
@@ -130,6 +135,13 @@ export const parse = (
   };
 };
 
+export const isStaticTemplateConfig = (
+  streamId: string | undefined,
+  streamConfig: StreamInternal | undefined
+): boolean => {
+  return !streamId && (!streamConfig || !streamConfig.$id);
+};
+
 export const convertTemplateModuleToTemplateModuleInternal = (
   templateFilepath: string,
   templateModule: TemplateModule<any, any>,
@@ -146,7 +158,7 @@ export const convertTemplateModuleToTemplateModuleInternal = (
     path: templateFilepath,
     filename: templatePath.base,
     templateName: templatePath.name,
-  };
+  } as TemplateModuleInternal<any, any>;
 
   validateTemplateModuleInternal(templateModuleInternal);
 
@@ -157,11 +169,16 @@ export const convertTemplateConfigToTemplateConfigInternal = (
   templateName: string,
   templateConfig: TemplateConfig | undefined
 ): TemplateConfigInternal => {
+  const stream = convertStreamToStreamInternal(templateConfig?.stream);
+
   return {
     name: templateConfig?.name ?? templateName,
     hydrate: templateConfig?.hydrate ?? true,
     ...templateConfig,
-    stream: convertStreamToStreamInternal(templateConfig?.stream),
+    stream: stream,
+    templateType: isStaticTemplateConfig(templateConfig?.streamId, stream)
+      ? "static"
+      : "entity",
   };
 };