+++ packages/pages/src/dev/server/middleware/indexPage.ts
@@ -105,7 +105,11 @@ export const indexPage =
           indexPageHtml = indexPageHtml.replace(
             `<!--static-pages-html-->`,
             `<h3>Static Pages</h3>
-            ${createStaticPageListItems(localDataManifest, devServerPort)}`
+            ${createStaticPageListItems(
+              localDataManifest,
+              useProdURLs,
+              devServerPort
+            )}`
           );
         }
 
@@ -187,47 +191,83 @@ export const indexPage =
     }
   };
 
+const formatStaticLink = (
+  useProdUrls: boolean,
+  templateName: string,
+  locale: string,
+  slug?: string
+) => {
+  if (useProdUrls) {
+    return slug;
+  }
+
+  // Default language is en. Only add a query param if not en.
+  const localeQuery = locale === "en" ? "" : `?locale=${locale}`;
+
+  return templateName + localeQuery;
+};
+
 const createStaticPageListItems = (
   localDataManifest: LocalDataManifest,
+  useProdURLs: boolean,
   devServerPort: number
 ) => {
   return Array.from(localDataManifest.static).reduce(
-    (templateAccumulator, [, { featureName, pathToLocaleMap }]) =>
-      templateAccumulator +
-      `<h4>${featureName} pages (${pathToLocaleMap.size}):</h4>` +
-      `<table>
-        <thead>
-          <tr>
-            <td>URL</td>
-            ${pathToLocaleMap.size > 1 ? "<td>Locale</td>" : ""}
-          </tr>
-        </thead>
-        <tbody>
-          ${[...pathToLocaleMap]
-            .map(
-              ([path, locale]) => `<tr>
-            ${
-              pathToLocaleMap.size > 1
-                ? `<td>
-                <a href="http://localhost:${devServerPort}/${path}?locale=${locale}">
-                  ${path}?locale=${locale}
-                </a>
-              </td>`
-                : `<td>
-                <a href="http://localhost:${devServerPort}/${path}">
-                  ${path}
-                </a>
-              </td>`
-            }
-            ${pathToLocaleMap.size > 1 ? `<td>${locale}</td>` : ""}`
-            )
-            .join("")}
-          </tr>
-        </tbody>
-      </table>
-    `,
+    (templateAccumulator, [, { featureName, pathToLocalesMap }]) => {
+      return (
+        templateAccumulator +
+        `<h4>${featureName} pages (${pathToLocalesMap.values.length}):</h4>` +
+        `<table>
+          <thead>
+            <tr>
+              <td>URL</td>
+              <td>Locale</td>
+            </tr>
+          </thead>
+          <tbody>
+            ${[...pathToLocalesMap].map(([path, locales]) =>
+              locales.map((locale) => {
+                const link = formatStaticLink(
+                  useProdURLs,
+                  featureName,
+                  locale,
+                  path
+                );
+
+                return `
+                  <tr>
+                    <td>
+                      <a href="http://localhost:${devServerPort}/${link}">
+                        ${link}
+                      </a>
+                    </td>
+                    <td>${locale}</td>
+                  </tr>`;
+              })
+            )}
+          </tbody>
+        </table>`
+      );
+    },
     ""
   );
+};
+
+const formatEntityLink = (
+  useProdUrls: boolean,
+  templateName: string,
+  entityId: string,
+  locale: string,
+  slug?: string
+) => {
+  if (useProdUrls) {
+    return slug;
+  }
+
+  // Default language is en. Only add a query param if not en.
+  const localeQuery = locale === "en" ? "" : `?locale=${locale}`;
+
+  return `${templateName}/${entityId}${localeQuery}`;
 };
 
 const createEntityPageListItems = (
@@ -236,16 +276,6 @@ const createEntityPageListItems = (
   useProdURLs: boolean,
   devServerPort: number
 ) => {
-  const formatLink = (entityId: string, slug: string | undefined) => {
-    if (useProdURLs) {
-      return `http://localhost:${devServerPort}/${slug}`;
-    }
-
-    return `http://localhost:${devServerPort}/${encodeURIComponent(
-      templateName
-    )}/${entityId}`;
-  };
-
   const { accountId, universe } = parseYextrcContents();
   const partition = getPartition(Number(accountId));
   // Content is knowledge graph
@@ -260,19 +290,29 @@ const createEntityPageListItems = (
   };
 
   const entities = localDataManifest.entity.get(templateName) || [];
-  return entities.reduce((entityAccumulator, { uid, entityId, slug }) => {
-    if (useProdURLs && !slug) {
-      logWarning(
-        `No document.slug found for entityId "${entityId}", no link will be rendered in the index page.`
+  return entities.reduce(
+    (entityAccumulator, { uid, entityId, slug, locale }) => {
+      if (useProdURLs && !slug) {
+        logWarning(
+          `No document.slug found for entityId "${entityId}", no link will be rendered in the index page.`
+        );
+        return entityAccumulator;
+      }
+
+      const link = formatEntityLink(
+        useProdURLs,
+        templateName,
+        entityId,
+        locale,
+        slug
       );
-      return entityAccumulator;
-    }
-    return (
-      entityAccumulator +
-      `<tr>
+
+      return (
+        entityAccumulator +
+        `<tr>
         <td>
-          <a href="${formatLink(entityId, slug)}">
-            ${slug ?? entityId}
+          <a href="http://localhost:${devServerPort}/${link}">
+            ${link}
            </a>
         </td>
         <td>
@@ -283,8 +323,10 @@ const createEntityPageListItems = (
           }
         </td>
     </tr>`
-    );
-  }, "");
+      );
+    },
+    ""
+  );
 };
 
 const getInfoMessage = (isDynamic: boolean, isProdUrl: boolean): string => {