+++ packages/pages/src/dev/server/server.ts
@@ -1,6 +1,5 @@
 import express from "express";
 import path from "path";
-import fs from "fs";
 import { createServer as createViteServer } from "vite";
 import { serverRenderRoute } from "./middleware/serverRenderRoute.js";
 import { ignoreFavicon } from "./middleware/ignoreFavicon.js";
@@ -19,9 +18,7 @@ import { convertFunctionModuleToFunctionModuleInternal } from "../../common/src/
 import { loadViteModule } from "./ssr/loadViteModule.js";
 import { FunctionModule } from "../../common/src/function/types.js";
 import { getViteServerConfig } from "../../common/src/loader/vite.js";
-import { logWarning } from "../../util/logError.js";
-import { isUsingConfig } from "../../util/config.js";
-import yaml from "yaml";
+
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
@@ -42,8 +39,6 @@ export const createServer = async (
 
   // initialize the default project structure and use to help configure the dev server
   const projectStructure = await ProjectStructure.init({ scope });
-
-  const defaultLocale = getDefaultLocale(projectStructure);
 
   // create vite using ssr mode
   const vite = await createViteServer(getViteServerConfig(projectStructure));
@@ -155,13 +150,11 @@ export const createServer = async (
           vite,
           dynamicGenerateData,
           projectStructure,
-          defaultLocale,
         })
       : serverRenderRoute({
           vite,
           dynamicGenerateData,
           projectStructure,
-          defaultLocale,
         })
   );
 
@@ -182,44 +175,4 @@ export const createServer = async (
   app.listen(devServerPort, () =>
     process.stdout.write(`listening on :${devServerPort}\n`)
   );
-};
-
-/**
- * Read features.json and set the default locale to the first locale listed
- * Default to en if features.json cannot be read or there is no locales entry
- */
-const getDefaultLocale = (projectStructure: ProjectStructure) => {
-  // Read features.json or config.yaml and set the default locale to the first locale listed.
-  // Default to en if features.json cannot be read or there is no locales entry.
-  try {
-    let parsedConfigfile;
-    const { config } = projectStructure;
-    if (isUsingConfig(config.rootFiles.config, config.scope)) {
-      parsedConfigfile = yaml.parseDocument(
-        fs.readFileSync(
-          path.join(config.scope ?? "", config.rootFiles.config),
-          "utf-8"
-        )
-      );
-    } else {
-      parsedConfigfile = JSON.parse(
-        fs.readFileSync(
-          path.resolve(
-            projectStructure.getSitesConfigPath().path,
-            projectStructure.config.sitesConfigFiles.features
-          ),
-          "utf-8"
-        )
-      );
-    }
-    if (parsedConfigfile.locales && parsedConfigfile.locales.length > 0) {
-      return parsedConfigfile.locales[0];
-    }
-  } catch (e: any) {
-    if (e !== "Error: ENOENT: no such file or directory") {
-      logWarning(e);
-    }
-  }
-
-  return "en";
 };